LPD-27495 Rename package
